Split autocompletion code in a separate file
