+++ librz/core/Makefile
@@ -8,7 +8,7 @@ RZ_DEPS+=rz_reg rz_search rz_syscall rz_socket rz_magic rz_crypto
 
 OBJS=core.o cmd.o cfile.o cconfig.o visual.o cio.o yank.o libs.o agraph.o
 OBJS+=fortune.o hack.o vasm.o patch.o cbin.o rtr.o cmd_api.o cmd_descs.o
-OBJS+=carg.o canal.o project.o gdiff.o casm.o disasm.o cplugin.o
+OBJS+=carg.o canal.o cautocmpl.o project.o gdiff.o casm.o disasm.o cplugin.o
 OBJS+=vmenus.o vmenus_graph.o vmenus_zigns.o zdiff.o citem.o
 OBJS+=task.o panels.o pseudo.o vmarks.o anal_tp.o anal_objc.o blaze.o
 OBJS+=cannotated_code.o serialize_core.o
